Trait tower_balance::choose::Choose

pub trait Choose<K, N> {
    fn choose(&mut self, replicas: Replicas<K, N>) -> usize;
}

A strategy for choosing nodes.

Required Methods

fn choose(&mut self, replicas: Replicas<K, N>) -> usize

Returns the index of a replica to be used next.

replicas cannot be empty, so this function must always return a valid index on [0, replicas.len()-1].