Function tower_balance::power_of_two_choices

pub fn power_of_two_choices<D, R>(
    loaded: D, 
    rng: R
) -> Balance<D, PowerOfTwoChoices<R>> where
    D: Discover,
    D::Service: Load,
    <D::Service as Load>::Metric: PartialOrd,
    R: Rng, 

Chooses services using the Power of Two Choices.

This configuration is prefered when a load metric is known.

As described in the Finagle Guide:

The algorithm randomly picks two services from the set of ready endpoints and selects the least loaded of the two. By repeatedly using this strategy, we can expect a manageable upper bound on the maximum load of any server.

The maximum load variance between any two servers is bound by ln(ln(n)) where n is the number of servers in the cluster.