Module tower_balance::p2c

A Power-of-Two-Choices Load Balancer

Structs

Balance
Distributes requests across inner services using the Power of Two Choices.

BalanceLayer
Efficiently distributes requests across an arbitrary number of services

BalanceMake
Makes Balancers given an inner service that makes Discovers.

MakeFuture
Makes a balancer instance.