Struct tower_balance::Balance

pub struct Balance<D: Discover, C> { /* fields omitted */ }

Balances requests across a set of inner services.

Methods

impl<D, C> Balance<D, C> where
D: Discover,
C: Choose<D::Key, D::Service>,

pub fn new(discover: D, choose: C) -> Self

Creates a new balancer.

pub fn is_not_ready(&self) -> bool

Returns true iff there are no ready services.

This is not authoritative and is only useful after poll_ready has been called.

pub fn num_ready(&self) -> usize

Counts the number of services considered to be ready.

This is not authoritative and is only useful after poll_ready has been called.

pub fn num_not_ready(&self) -> usize

Counts the number of services not considered to be ready.

This is not authoritative and is only useful after poll_ready has been called.