Struct tower_balance::choose::PowerOfTwoChoices

pub struct PowerOfTwoChoices<R> { /* fields omitted */ }

Chooses nodes using the Power of Two Choices.

This is a load-aware strategy, so this may only be used to choose over services that implement Load.

As described in the Finagle Guide:

The algorithm randomly picks two nodes from the set of ready endpoints and selects the least loaded of the two. By repeatedly using this strategy, we can expect a manageable upper bound on the maximum load of any server.

The maximum load variance between any two servers is bound by ln(ln(n)) where n is the number of servers in the cluster.

Methods

impl<R: Rng> PowerOfTwoChoices<R>

pub fn new(rng: R) -> Self