Struct tower_balance::Balance

pub struct Balance<D: Discover, C> { /* fields omitted */ }

Balances requests across a set of inner services.

Methods

impl<D> Balance<D, PowerOfTwoChoices> where
D: Discover,
D::Service: Load,
<D::Service as Load>::Metric: PartialOrd + Debug,

pub fn p2c(discover: D) -> Self

Chooses services using the Power of Two Choices.

This configuration is prefered when a load metric is known.

As described in the Finagle Guide:

The algorithm randomly picks two services from the set of ready endpoints and selects the least loaded of the two. By repeatedly using this strategy, we can expect a manageable upper bound on the maximum load of any server.

The maximum load variance between any two servers is bound by ln(ln(n)) where n is the number of servers in the cluster.

pub fn p2c_with_rng<R: Rng>(discover: D, rng: &mut R) -> Result<Self, Error>

Initializes a P2C load balancer from the provided randomization source.

This may be preferable when an application instantiates many balancers.

impl<D: Discover> Balance<D, RoundRobin>

pub fn round_robin(discover: D) -> Self

Attempts to choose services sequentially.

This configuration is prefered when no load metric is known.

impl<D, C> Balance<D, C> where
D: Discover,
C: Choose<D::Key, D::Service>,

pub fn new(discover: D, choose: C) -> Self

Creates a new balancer.

pub fn is_ready(&self) -> bool

Returns true iff there are ready services.

This is not authoritative and is only useful after poll_ready has been called.

pub fn is_not_ready(&self) -> bool

Returns true iff there are no ready services.

This is not authoritative and is only useful after poll_ready has been called.

pub fn num_ready(&self) -> usize

Counts the number of services considered to be ready.

This is not authoritative and is only useful after poll_ready has been called.

pub fn num_not_ready(&self) -> usize

Counts the number of services not considered to be ready.

This is not authoritative and is only useful after poll_ready has been called.